can delete and edit tasks

+++ app/scripts/main.js
@@ -128,18 +128,19 @@ var Router = Backbone.Router.extend({
       });
 
       $('#delete-project').on('click', function(){
-        App.deleteProject();
+        var result = confirm("Do you want to delete this project?");
+        if (result) {
+          App.deleteProject();
+        }
         trace('hi delete button is active');
       });
 
       $('#show-members').on('click', function(){
         App.showMembers(projectId);
-        trace('hi show members button is active');
       });
 
       $('#new-task').on('click', function(){
         Task.newTask();
-        trace('hi new task button is active <:o');
       });
 
     }).fail(function(jqXHR, textStatus, errorThrown){
@@ -180,9 +181,26 @@ var Router = Backbone.Router.extend({
         trace('hi new comment button is active <:o');
       });
 
+      $('#update-task').on('click', function(){
+        Task.updateTask();
+        trace('hi update button is active');
+      });
+
       $('#show-comment').on('click', function(){
         Comment.showComments(taskId);
         trace('hi show comment button is active <:o');
+      });
+
+      $('.delete-comment').on('click', function(){
+        Comment.deleteComment(taskId);
+        trace('hi show comment button is active <:o');
+      });
+
+      $('#delete-task').on('click', function(){
+        var result = confirm("Do you want to delete this task?");
+        if (result) {
+          Task.deleteTask();
+        }
       });
 
     }).fail(function(jqXHR, textStatus, errorThrown){
@@ -299,61 +317,80 @@ Task.newTaskParams = function(name, date, status, order, privacy, router){
 };
 
 Task.updateTask = function(){
-  trace('hello from the update project backbone');
-    $('#container').empty().load('partials/project-form.html',function(response,status,xhr){
-      var $form = $('#project-form');
+  trace('hello from the update task backbone');
+    $('#container').empty().load('partials/task-form.html',function(response,status,xhr){
+      var $form = $('#task-form');
       $form.on('submit',function(event){
-        Project.updateProcessForm(event,$form);
+        Task.updateTaskForm(event,$form);
       });
     });
 };
 
-Task.updateProcessForm = function(e,form){
+Task.updateTaskForm = function(e,form){
   trace('hello from the project update form');
   var locate = window.location.hash;
   var point = locate.lastIndexOf('/');
-  var projectId = parseInt(locate.substring(point+1, locate.length));
+  var taskId = parseInt(locate.substring(point+1, locate.length));
   if(e.preventDefault) e.preventDefault();
-  var name = $(form).find("input[name='proj-name']").val();
-  var description = $(form).find("input[name='proj-description']").val();
-  var date = $(form).find("input[name='proj-date']").val();
-  var privacy = $(form).find("select[name='proj-privacy']").val();
-  var user = localStorage.getItem('currentUser');
-  Project.updatePostParams(name, description, date, privacy, user, projectId);
+  var name = $(form).find("input[name='task-name']").val();
+  var date = $(form).find("input[name='task-date']").val();
+  var status = $(form).find("select[name='task-status']").val();
+  var order = $(form).find("select[name='task-order']").val();
+  var privacy = $(form).find("select[name='task-privacy']").val();
+  Task.updateTaskParams(name, date, status, order, privacy, taskId);
 };
 
-Task.updatePostParams = function(name, description, date, privacy, user, projectId){
+Task.updateTaskParams = function(name, date, status, order, privacy, taskId){
   $.ajax({
-    url: App.url + '/projects/' + projectId,
+    url: App.url + '/tasks/' + taskId,
     type: 'PATCH',
     data: {
-      project: {
+      task: {
         name: name,
-        description: description,
         due_date: date,
+        status: status,
+        order: order,
         privacy: privacy
       },
     },
     complete: function(jqXHR,textStatus){
-      trace(jqXHR, textStatus, "complete post!!");
+      trace(jqXHR, textStatus, "complete task!!");
     },
     success: function(data, textStatus, jqXHR){
       router.navigate("projects",{trigger: true});
-      trace(data,textStatus, jqXHR, "successful post!!");
+      trace(data,textStatus, jqXHR, "successful task!!");
     },
     error: function(jqXHR,error,exception){
       trace(jqXHR,error,exception);
     },
   }).done(function(response){
-    trace(response, "posted project!!");
+    trace(response, "posted task!!");
   }).fail(function(jqXHR, textStatus, thrownError){
     trace(jqXHR, textStatus, thrownError);
-    router.navigate("projects",{trigger: true});
+    router.navigate("tasks",{trigger: true});
     trace('wat');
   }).always(function(response){
     trace(response);
   });
 };
+
+Task.deleteTask = function(){
+  $('#container').empty();
+  $('.jumbotron').hide();
+  var locate = window.location.hash;
+  var point = locate.lastIndexOf('/');
+  var taskId = parseInt(locate.substring(point+1, locate.length));
+  $.ajax({
+    url: App.url + '/tasks/' + taskId,
+    type: 'DELETE',
+  }).done(function(data){
+    trace(data);
+    trace('deleted project');
+    window.location.href = '/#/projects';
+  }).fail(function(jqXHR, textStatus, errorThrown){
+    trace(jqXHR, textStatus, errorThrown);
+  });
+}
 
 Comment.newComment = function(){
   trace('hello from the new comment backbone');
@@ -430,6 +467,24 @@ Comment.showComments = function(taskId){
     trace(response);
   });
 };
+
+Comment.deleteComment = function(taskId){
+  $('#container').empty();
+  $('.jumbotron').hide();
+  // var locate = window.location.hash;
+  // var point = locate.lastIndexOf('/');
+  // var taskId = parseInt(locate.substring(point+1, locate.length));
+  $.ajax({
+    url: App.url + '/tasks/' + taskId,
+    type: 'DELETE',
+  }).done(function(data){
+    trace(data);
+    trace('deleted project');
+    window.location.href = '/#/projects';
+  }).fail(function(jqXHR, textStatus, errorThrown){
+    trace(jqXHR, textStatus, errorThrown);
+  });
+}
 
 App.updateProject = function(){
   trace('hello from the update project backbone');
